requirements/basic_requirements.txt
# Python packages
#-----------------

# http://code.google.com/p/feedparser/
# http://pypi.python.org/pypi/feedparser/
feedparser>=5.1.2,<5.2

# http://pygments.org/
# http://pypi.python.org/pypi/Pygments
Pygments>=1.5,<1.6

# http://trac.saddi.com/flup
# http://pypi.python.org/pypi/flup
# needed, for fastCGI usage
flup>=1.0.2,<1.1

# external Django apps
#----------------------

# https://github.com/etianen/django-reversion
# http://pypi.python.org/pypi/django-reversion
django-reversion>=1.6,<1.7

# https://github.com/jezdez/django-dbtemplates
# http://pypi.python.org/pypi/django-dbtemplates
django-dbtemplates>=1.3,<1.4

# https://github.com/jezdez/django_compressor
# http://pypi.python.org/pypi/django_compressor/
django_compressor==dev

# http://code.google.com/p/django-tagging/
# http://pypi.python.org/pypi/django-tagging
django-tagging>=0.3.1,<0.4

# http://south.aeracode.org/
# http://pypi.python.org/pypi/South
South>=0.7.5,<0.8
